feat: throw error if try to delete a product with sales

+++ services/product.service.js
@@ -1,8 +1,9 @@
 import ProductRepository from '../repositories/product.repository.js';
 import SupplierRepository from '../repositories/supplier.repository.js';
+import SaleRepository from '../repositories/sale.repository.js';
 
 async function createProduct(product) {
-    if (await SupplierRepository.getSupplier(product.supplier_id)) {
+    if (await SupplierRepository.getSupplier(product.supplierId)) {
         return await ProductRepository.insertProduct(product);
     }
     throw new Error('O supplier_id informado não existe.');
@@ -17,6 +18,12 @@ async function getProduct(id) {
 }
 
 async function deleteProduct(id) {
+    const sales = await SaleRepository.getSalesByProductId(id);
+    if (sales.length > 0) {
+        throw new Error(
+            'Não é possível excluir o produto pois ele tem vendas.'
+        );
+    }
     await ProductRepository.deleteProduct(id);
 }
